+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added upload_image_directory function for Dataset class (sc-3367)
+
 ### Changed
 
 - Upload local images asynchronously (sc-3366)